Add public method to reset internal manager caches
In particular, HTTP headers get added in `start` to the operation manager's `requestSerializer`, but then if those headers are cleared elsewhere in the application, the old headers will still be used for future requests. Clients can use this new `resetManagers` method to delete the old state after, for example, a user logs out.