Reuse AFHTTPSessionManagers if possible to avoid race condition crash
tl;dr: Using multiple AFHTTPSessionManagers for a given baseURL can result in a race condition involving overreleased SecTrustRefs that crash the application. This change attempts to re-use those managers as much as possible, and will hopefully fix the crash.

-----

It seems that the NSURLAuthenticationChallenge objects passed to NSURLSessionDelegate’s URLSession:didReceiveChallenge:completionHandler: share their protectionSpace.serverTrust SecTrustRefs. This is perhaps a bug in NSURLSession, but if they do share SecTrustRefs, then there's [an old Apple Listserv post](http://lists.apple.com/archives/apple-cdsa/2010/Apr/msg00017.html) that suggests it may cause a certain class of crashes related to CFNetworkTrust::evaluate().

By creating a new AFHTTPSessionManager every time a request is made to a particular baseURL, there might sometimes be multiple concurrent callbacks to NSURLSessionDelegate's URLSession:didReceiveChallenge:completionHandler, which the managers implement. From there, the challenge.protectionSpace.serverTrust [is passed](https://github.com/AFNetworking/AFNetworking/blob/3.1.0/AFNetworking/AFURLSessionManager.m#L978) to its AFSecurityPolicy when applicable, and then later, the SecTrustRef is evaluated and eventually released ([here](https://github.com/AFNetworking/AFNetworking/blob/3.1.0/AFNetworking/AFSecurityPolicy.m#L74) or [here](https://github.com/AFNetworking/AFNetworking/blob/3.1.0/AFNetworking/AFSecurityPolicy.m#L135)). If a SecTrustRef is shared between requests to a given baseURL, then there could be a race condition in which multiple simultaneous requests overrelease the SecTrustRef and a subsequent attempt to use it crashes the app.

-----

Note that the crash may still occur when using multiple AFHTTPSessionManagers with different requestTypes and responseTypes for a given baseURL. That seems significantly less likely to happen, however, so I wanted to see if this fixes the issue before attempting more significant changes.

It has to be in an @synchronized block because NSMutableDictionary is not thread safe, and because we don't want multiple managers created for a single key.

This may eventually be fixed by https://github.com/AFNetworking/AFNetworking/pull/3616, but this seems like a good change for SGHTTPRequest as well.