Use AFCompoundResponseSerializer to fix race condition
This fixes an issue that was introduced in https://github.com/seatgeek/SGHTTPRequest/pull/6.

With the prior implementation, multiple requests to a single baseURL would overwrite each other's responseSerializers, which could cause errors if they were expecting different content types. AFCompoundResponseSerializer avoids this problem, but we should instead use SGHTTPRequest's responseFormat to make sure that we still register an error if the response is not of the expected format.